begin moving propose to LightWallet from LightClient
