+++ zingolib/src/wallet.rs
@@ -55,6 +55,7 @@ pub mod utils;
 //these mods contain pieces of the impl LightWallet
 pub mod describe;
 pub mod disk;
+pub mod propose;
 pub mod send;
 pub mod witnesses;
